+++ wisply-chatbot/includes/class-admin.php
@@ -230,6 +230,7 @@ class Wisply_Admin {
             'emergency_msg_he', 'emergency_msg_en', 'emergency_msg_ru',
             'emergency_phone', 'emergency_eran_url', 'emergency_sahar_url',
             'report_recipients', 'report_daily', 'report_weekly',
+            'woo_enabled', 'woo_max_products', 'woo_show_stock', 'woo_visual_search',
         ];
         $secret_keys = [ 'ai_api_key', 'openai_api_key' ];
         $errors      = [];